Move tag to internal/tag
